+++ src/geometry/plane.rs
@@ -3,18 +3,19 @@ use ultraviolet::Vec3;
 
 use crate::geometry::{Aabb, AngularExt, Boxable, Intersectable, Intersection};
 use crate::geometry::ray::Ray;
+use crate::Float;
 
 /// A geometrical plane.
 #[derive(Debug, Default, Deserialize, Serialize)]
 pub struct Plane {
     /// The distance of the plane into the normal direction.
-    pub d: f32,
+    pub d: Float,
     /// The normal of a plane.
     pub normal: Vec3,
 }
 
 impl Plane {
-    pub fn new(d: f32, normal: Vec3) -> Self {
+    pub fn new(d: Float, normal: Vec3) -> Self {
         Self {
             d,
             normal: normal.normalized(),
@@ -25,17 +26,17 @@ impl Plane {
     #[inline]
     #[must_use]
     pub fn project(&self, point: Vec3) -> Vec3 {
-        point - (self.normal * self.distance(point))
+        point - (self.normal * self.distance(point) as f32)
     }
 
     /// Calculates the distance of the given point to the plane.
     #[inline]
     #[must_use]
-    pub fn distance(&self, point: Vec3) -> f32 {
-        let v = point - (self.normal * self.d);
+    pub fn distance(&self, point: Vec3) -> Float {
+        let v = point - (self.normal * self.d as f32);
         let angle = self.normal.angle_to(v);
 
-        angle.cos() * v.mag()
+        angle.cos() * v.mag() as Float
     }
 }
 
@@ -45,7 +46,7 @@ impl From<(Vec3, Vec3)> for Plane {
         let normal = pos_normal.1.normalized();
 
         let angle = position.angle_to(normal);
-        let d = angle.cos() * position.mag();
+        let d = angle.cos() * position.mag() as Float;
 
         Self { d, normal }
     }
@@ -59,13 +60,13 @@ impl Boxable for Plane {
 
 impl Intersectable<Ray> for Plane {
     fn intersects(&self, ray: &Ray) -> Option<Intersection> {
-        let denom = self.normal.dot(ray.direction);
-        if denom.abs() <= f32::EPSILON {
+        let denom = self.normal.dot(ray.direction) as Float;
+        if denom.abs() <= Float::EPSILON {
             return None;
         }
 
-        let t = -(self.normal.dot(ray.origin) + self.d) / denom;
-        if t <= f32::EPSILON {
+        let t = -(self.normal.dot(ray.origin) as Float + self.d) / denom;
+        if t <= Float::EPSILON {
             return None;
         }
